run_qemu: Add dcd tests
DCD requires additional testing beyond cxl_test.  Add a script which
will run on the host that can use QMP to inject/remove extents while
also running commands on the guest to monitor the reaction of those
extents.

The basic tests are the same as those done with cxl_test.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>